# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Farmer Registration Pilot'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://farmer-pilot.test'
paths:
...
  /api/farmer/register:
    post:
      summary: 'Register Farmer'
      operationId: registerFarmer
      description: 'This endpoint allows a user to register a farmer'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'required The authorization token. Example: Bearer {token}'
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"status\": \"success\",\n\"message\": \"Farmer profile created successfully\",\n\"data\": {\n\"first_name\": \"John\",\n\"last_name\": \"Doe\",\n\"dob\": \"1981-05-06\",\n}\n}"
        400:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Bad request'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Bad request'
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Unauthorized
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Unauthorized
        403:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: Forbidden
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: Forbidden
        404:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Not found'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Not found'
        405:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Method not allowed'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Method not allowed'
        422:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Validation error'
                  errors: {  }
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Validation error'
                  errors:
                    type: object
                    properties: []
        429:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Too many requests'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Too many requests'
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  status: error
                  message: 'Server error'
                properties:
                  status:
                    type: string
                    example: error
                  message:
                    type: string
                    example: 'Server error'
      tags:
        - 'Farmer Profile'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  description: 'The first name of the farmer.'
                  example: John
                last_name:
                  type: string
                  description: 'The last name of the farmer.'
                  example: Doe
                dob:
                  type: date
                  description: 'The date of birth of the farmer.'
                  example: '1990-01-01'
                gender:
                  type: string
                  description: 'Farmer gender. Example Male/Female'
                  example: eum
                education_level:
                  type: string
                  description: 'The education level of the farmer.'
                  example: Primary
                phone_number:
                  type: string
                  description: 'The phone number of the farmer.'
                  example: '0789123456'
                id_number:
                  type: string
                  description: 'The national ID number of the farmer.'
                  example: CM12345678
                marital_status:
                  type: string
                  description: 'The marital status of the farmer.'
                  example: Married
                district:
                  type: string
                  description: 'The district of the farmer.'
                  example: Kampala
                county:
                  type: string
                  description: 'The county of the farmer.'
                  example: Makindye
                sub_county:
                  type: string
                  description: 'The sub county of the farmer.'
                  example: Kibuye
                parish:
                  type: string
                  description: 'The parish of the farmer.'
                  example: Kibuye
                village:
                  type: string
                  description: 'The village of the farmer.'
                  example: Kibuye
                fpo_id:
                  type: integer
                  description: 'The FPO name of the farmer.'
                  example: 0
                farmer_cordinates:
                  type: string
                  description: 'The farmer cordinates of the farmer.'
                  example: '0.0000,0.0000'
                next_of_kin:
                  type: string
                  description: 'The next of kin of the farmer.'
                  example: 'Jane Doe'
                next_of_kin_contact:
                  type: string
                  description: 'The next of kin contact of the farmer.'
                  example: '0789123456'
                next_of_kin_relationship:
                  type: string
                  description: 'The next of kin relationship of the farmer.'
                  example: Wife
                male_members_in_household:
                  type: integer
                  description: 'Number of male members of the household.'
                  example: 2
                female_members_in_household:
                  type: integer
                  description: 'Number of male members of the household.'
                  example: 4
                members_above_18:
                  type: integer
                  description: 'Number of members above 18 years of age.'
                  example: 4
                children_below_5:
                  type: integer
                  description: 'Number of children below 5 years of age.'
                  example: 2
                school_going_children:
                  type: integer
                  description: 'Number of school going children.'
                  example: 2
                head_of_family:
                  type: string
                  description: 'The head of family of the farmer.'
                  example: 'John Doe'
                how_much_do_you_earn_from_agricultural_activities:
                  type: string
                  description: 'How much do you earn from agricultural activities.'
                  example: '100000'
                how_much_do_you_earn_from_non_agricultural_activities:
                  type: string
                  description: 'How much do you earn from non agricultural activities.'
                  example: '100000'
                do_you_have_an_account_with_an_FI:
                  type: string
                  description: 'Do you have an account with an FI.'
                  example: 'Yes'
                farm_size:
                  type: string
                  description: 'The farm size of the farmer.'
                  example: '2'
                farm_size_under_agriculture:
                  type: string
                  description: 'The farm size under agriculture of the farmer.'
                  example: '2'
                land_ownership:
                  type: string
                  description: 'The land ownership of the farmer.'
                  example: Owned
                type_of_farming:
                  type: string
                  description: 'The type of farming of the farmer.'
                  example: Crop
                crops_grown:
                  type: string
                  description: 'The crops grown of the farmer.'
                  example: Maize
                animals_kept:
                  type: string
                  description: 'The animals kept of the farmer.'
                  example: Cattle
                estimated_produce_value_last_season:
                  type: string
                  description: 'The estimated produce value last season of the farmer.'
                  example: '100000'
                estimated_produce_value_this_season:
                  type: string
                  description: 'The estimated produce value this season of the farmer.'
                  example: '100000'
                data_captured_by:
                  type: string
                  description: 'The data captured by of the farmer.'
                  example: '1'
                agent_id:
                  type: integer
                  description: 'The agent id of the farmer.'
                  example: 1
              required:
                - first_name
                - last_name
                - dob
                - education_level
                - phone_number
                - id_number
                - marital_status
                - district
                - county
                - sub_county
                - parish
                - village
                - fpo_id
                - farmer_cordinates
                - next_of_kin
                - next_of_kin_contact
                - next_of_kin_relationship
                - male_members_in_household
                - female_members_in_household
                - members_above_18
                - children_below_5
                - school_going_children
                - head_of_family
                - how_much_do_you_earn_from_agricultural_activities
                - how_much_do_you_earn_from_non_agricultural_activities
                - do_you_have_an_account_with_an_FI
                - farm_size
                - farm_size_under_agriculture
                - land_ownership
                - type_of_farming
                - crops_grown
                - animals_kept
                - estimated_produce_value_last_season
                - estimated_produce_value_this_season
                - data_captured_by
                - agent_id
...